Домашна 3/technical-analysis/oscilators/requirements.txt
numpy==2.3.5
pandas==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.44
pandas_ta==0.3.14b0
tqdm==4.66.5